:truck:Modify : format-time 내부 함수들 function.ts에 이동
