+++ lib/function.ts
@@ -2,6 +2,8 @@ import { create } from "zustand"
 import { MoodType } from "./mock-data"
 import { ApiResSuccess } from "@/types/api-response"
 import { DiaryTypes } from "@/types/my-diarys"
+import { parse, format } from "date-fns"
+import { ko } from "date-fns/locale"
 
 const moodImgData: MoodType = {
   happy: "/assets/svg/happy.svg",
@@ -63,4 +65,16 @@ export const sortDiarys = (diarys: DiaryTypes[]) => {
     },
     {}
   )
+}
+
+export const getKoTime = (val: string) => {
+  const date = parse(val, "yyyy.MM.dd HH:mm:ss", new Date())
+  const formattedTime = format(date, "aa h:mm", { locale: ko })
+  return formattedTime
+}
+
+export const getKoDate = (val: string) => {
+  const date = parse(val, "yyyy.MM.dd HH:mm:ss", new Date())
+  const formattedDate = format(date, `M월 d일 E요일`, { locale: ko })
+  return formattedDate
 }
